+++ django/applications/catmaid/static/js/widgets/system-check.js
@@ -30,8 +30,15 @@
     return {
       controlsID: 'system-widget-controls-' + this.widgetID,
       contentID: 'system-widget-' + this.widgetID,
+      class: "system-check",
       createControls: function(controls) {
         var self = this;
+
+        CATMAID.DOM.appendButton(controls,
+            "Update", "Update server stats",
+            function() {
+              self.update();
+            });
 
         CATMAID.DOM.appendButton(controls,
             "Run tests", "Run various system performance tests",
@@ -94,6 +101,18 @@
         content.appendChild(fpsHeader);
         this.fpsResult = document.createElement('p');
         content.appendChild(this.fpsResult);
+
+        if (CATMAID.hasPermission(project.id, 'can_administer')) {
+          let serverHeader = content.appendChild(document.createElement('h2'));
+          serverHeader.appendChild(document.createTextNode("Server"));
+          this.serverResults = content.appendChild(document.createElement('div'));
+          let dbHeader = content.appendChild(document.createElement('h2'));
+          dbHeader.appendChild(document.createTextNode("Database"));
+          this.dbResults = content.appendChild(document.createElement('div'));
+        }
+      },
+      init: function() {
+        this.update();
       }
     };
   };
@@ -147,6 +166,211 @@
 
 
 		increment();
+  };
+
+  SystemCheckWidget.prototype.update = function() {
+    if (CATMAID.hasPermission(project.id, 'can_administer')) {
+      CATMAID.fetch(project.id + '/stats/server')
+        .then(stats => {
+          this.redrawServerStats(stats['server']);
+          this.redrawDatabaseStats(stats['database']);
+        });
+    }
+  };
+
+  let serverStats = {
+    'load_avg': {
+      name: 'Load average',
+      comment: 'Processes that want to run averaged over last 1, 5 and 15 min.',
+      render: function(value) {
+        return value.join(', ');
+      },
+    },
+  };
+
+  SystemCheckWidget.prototype.redrawServerStats = function(stats) {
+    // Clear server stats
+    while (this.serverResults.lastChild) {
+      this.serverResults.removeChild(this.serverResults.lastChild);
+    }
+
+    if (!stats || CATMAID.tools.isEmpty(stats)) {
+      this.serverResults.innerHTML = '<p><em>No data found</em></p>';
+    }
+
+    let table = document.createElement('table');
+    table.appendChild(document.createElement('th'));
+    table.appendChild(document.createElement('tbody'));
+    this.serverResults.appendChild(table);
+
+    let statsObjects = Object.keys(stats).map(statName => ({
+      'name': statName,
+      'value': stats[statName],
+    }));
+
+    let datatable = $(table).DataTable({
+      data: statsObjects,
+      dom: "lrfhtip",
+      order: [],
+      autoWidth: false,
+      paging: true,
+      lengthMenu: [CATMAID.pageLengthOptions, CATMAID.pageLengthLabels],
+      columns: [{
+        title: 'Name',
+        data: 'name',
+        width: '10em',
+        class: 'cm-center',
+        render: function(data, type, row, meta) {
+          let stat = serverStats[data];
+          return stat ? stat.name : data;
+        }
+      }, {
+        title: 'Value',
+        data: 'value',
+        width: '10em',
+        class: 'cm-center',
+        render: function(data, type, row, meta) {
+          let stat = serverStats[row.name];
+          if (stat && CATMAID.tools.isFn(stat.render)) {
+            return stat.render(data);
+          }
+          return data;
+        }
+      }, {
+        title: 'Comment',
+        render: function(data, type, row, meta) {
+          let stat = serverStats[row.name];
+          return stat ? (stat.comment || '') : '';
+        }
+      }],
+    });
+  };
+
+  /**
+   * Database stats. Most suggestions are taken as rough rule of thumb from the
+   * talk "Deep Dive into Postgres Statistics" by Alexey Lesovsky.
+   */
+  let databaseStats = {
+    'version': {
+      name: 'PostgreSQL Version',
+      comment: 'CATMAID requires at least version 10 of Postgres.',
+    },
+    'c_ratio': {
+      name: 'Committed transactions ',
+      comment: 'Percentage of committed versus rolled back transactions. Should be > 95%.',
+      render: value => `${value}%`,
+    },
+    'deadlocks': {
+      name: 'Deadlocks',
+      comment: 'Number of deadlocks in system. Should be zero.',
+    },
+    'conflicts': {
+      name: 'Conflicts',
+      comment: 'Number of queries canceled due to conflicts with recovery. Should be < 10.',
+    },
+    'temp_files': {
+      name: 'Temporary files',
+      comment: 'Number of temporary files created by queries. Should be < 100.',
+    },
+    'temp_size': {
+      name: 'Temp file size',
+      comment: 'Total data written to temporary files. Should be below 10 GB.',
+    },
+    'blks_read': {
+      name: 'Blocks read',
+      comment: 'Number of disk blocks read in the database.',
+    },
+    'blks_hit': {
+      name: 'Cached blocks read',
+      comment: 'Number of times blocks were already found in buffer cache.',
+    },
+    'cache_hit_ratio': {
+      name: 'Cache hit ratio',
+      comment: 'Indicates what percentage of data reads hit the cache. Should be > 95%',
+      render: (value) => `${Number(value).toFixed(2)}%`,
+    },
+    'checkpoints_req': {
+      name: 'Requested checkpoins',
+      comment: 'Number of times checkpoints have been requeste. Should be < timed checkpoints.',
+    },
+    'checkpoints_timed': {
+      name: 'Timed checkpoins',
+      comment: 'Number of scheduled checkpoints have been performed. Should be > requested checkpoints.',
+    },
+    'buffers_clean': {
+      name: 'Writter buffers',
+      comment: 'Numbr of buffers written by background writer',
+    },
+    'maxwritten_clean': {
+      name: 'Canceled buffer writes',
+      comment: 'Number of times buffer writing was stopped, because too many buffers were written. Should be zero.',
+    },
+    'buffers_backend_fsync': {
+      name: 'Own fsync calls by backend',
+      comment: 'Number of times a backend had to to execute its own fsync call. Should be zero.',
+    },
+    'replication_lag': {
+      name: 'Replication lag (sec)',
+      comment: 'If replication is enabled, this shows how far behind replicas are (in seconds). Should be low.',
+    }
+  };
+
+  SystemCheckWidget.prototype.redrawDatabaseStats = function(stats) {
+    // Clear database stats
+    while (this.dbResults.lastChild) {
+      this.dbResults.removeChild(this.dbResults.lastChild);
+    }
+
+    if (!stats || CATMAID.tools.isEmpty(stats)) {
+      this.dbResults.innerHTML = '<p><em>No data found</em></p>';
+    }
+
+    let table = document.createElement('table');
+    table.appendChild(document.createElement('th'));
+    table.appendChild(document.createElement('tbody'));
+    this.dbResults.appendChild(table);
+
+    let statsObjects = Object.keys(stats).map(statName => ({
+      'name': statName,
+      'value': stats[statName],
+    }));
+
+    let datatable = $(table).DataTable({
+      data: statsObjects,
+      dom: "lrfhtip",
+      order: [],
+      autoWidth: false,
+      paging: true,
+      lengthMenu: [CATMAID.pageLengthOptions, CATMAID.pageLengthLabels],
+      columns: [{
+        title: 'Name',
+        data: 'name',
+        width: '11em',
+        class: 'cm-center',
+        render: function(data, type, row, meta) {
+          let stat = databaseStats[data];
+          return stat ? stat.name : data;
+        }
+      }, {
+        title: 'Value',
+        data: 'value',
+        width: '16em',
+        class: 'cm-center',
+        render: function(data, type, row, meta) {
+          let stat = databaseStats[row.name];
+          if (stat && CATMAID.tools.isFn(stat.render)) {
+            return stat.render(data);
+          }
+          return data;
+        }
+      }, {
+        title: 'Comment',
+        render: function(data, type, row, meta) {
+          let stat = databaseStats[row.name];
+          return stat ? (stat.comment || '') : '';
+        }
+      }],
+    });
   };
 
   CATMAID.SystemCheckWidget = SystemCheckWidget;